Add Channel CRUD

+++ server/cmd/api/helpers.go
@@ -12,13 +12,13 @@ import (
 	"github.com/go-chi/chi/v5"
 )
 
-func (app *application) readIDParam(r *http.Request) (int, error) {
+func (app *application) readIDParam(r *http.Request) (int64, error) {
 	id, err := strconv.Atoi(chi.URLParamFromCtx(r.Context(), "id"))
 	if err != nil || id < 1 {
 		return 0, errors.New("invalid id parameter")
 	}
 
-	return id, nil
+	return int64(id), nil
 }
 
 type envelope map[string]interface{}
